zcash_pool_migration_sqlite: split the row (de)serialization into a codec module
Extract the pure, SQL-agnostic byte (de)serialization of the engine types
out of the store into a crate-internal `codec` module, and the `Error`
type into an `error` module, so `store` holds only the DDL and the SQL.
Add a `Writer` cursor mirroring the existing `Reader`, so encoding is
symmetric with decoding. No format or public-API change: the on-disk
bytes, the SQLite column set, and the `orchard_ironwood` public surface
are unchanged.
